Improved Box and Order classes
Box class has all properties of a box as well as a method setter to set
the type of box.
Order class now has a method to check the type of box and set it as well
as a test method which makes one Box object with test data and prints it
on program start.